FE refactoring finished
